requirements.txt
numpy
chromadb
pandas
pypdf
python-dotenv
langchain
langchain-community
langchain-text-splitters
umap-learn
matplotlib
sentence-transformers
transformers
torch
huggingface-hub
ollama
